Now installation tests skip sfst or foma if they are not found.
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@3021 941e2c2b-deac-454f-805a-451daa25f33c

+++ swig/test/test_std_streams.sh
@@ -7,7 +7,7 @@
 #
 
 # Check that these tools are found
-for tool in hfst-strings2fst hfst-compare;
+for tool in hfst-strings2fst hfst-compare hfst-format;
 do
     if (which $tool | grep "not found" ); then
 	echo "ERROR: the tool "$tool" was not found, install it before running this test (test_std_streams.sh)"
@@ -30,36 +30,46 @@ fi
 STRING1="abcdefghijklmnopqrstuvwxyz"
 STRING2="ABCDEFGHIJKLMNOPQRSTUVWXYZ"
 
-echo $STRING1 | $STRINGS2FST -f sfst > tr1.sfst
-echo $STR1NG2 | $STRINGS2FST -f sfst > tr2.sfst
-cat tr1.sfst tr2.sfst > tr.sfst
-rm tr1.sfst tr2.sfst
+if (hfst-format --list | grep 'sfst' > /dev/null); then
+    echo $STRING1 | $STRINGS2FST -f sfst > tr1.sfst
+    echo $STR1NG2 | $STRINGS2FST -f sfst > tr2.sfst
+    cat tr1.sfst tr2.sfst > tr.sfst
+    rm tr1.sfst tr2.sfst
+else
+    echo "warning: sfst back-end not available, assumed switched off and skipping"
+fi
 
 echo $STRING1 | $STRINGS2FST -f openfst-tropical > tr1.ofst
 echo $STRING2 | $STRINGS2FST -f openfst-tropical > tr2.ofst
 cat tr1.ofst tr2.ofst > tr.ofst
 rm tr1.ofst tr2.ofst
 
-echo $STRING1 | $STRINGS2FST -f foma > tr1.foma
-echo $STRING2 | $STRINGS2FST -f foma > tr2.foma
-cat tr1.foma tr2.foma > tr.foma
-rm tr1.foma tr2.foma
+if (hfst-format --list | grep 'foma' > /dev/null); then
+    echo $STRING1 | $STRINGS2FST -f foma > tr1.foma
+    echo $STRING2 | $STRINGS2FST -f foma > tr2.foma
+    cat tr1.foma tr2.foma > tr.foma
+    rm tr1.foma tr2.foma
+else
+    echo "warning: foma back-end not available, assumed switched off and skipping"
+fi
 
 
 # Perform the tests for all transducer types:
 
 # SFST
-if ! (cat tr.sfst | $PYTHON ./test_std_streams.py sfst 2 > tmp.sfst); then
-    echo "ERROR: in sfst"
-    rm tr.sfst
-    exit 1
-fi
-if ! ($COMPARE tr.sfst tmp.sfst); then
-    echo "ERROR in sfst"
+if (hfst-format --list | grep 'sfst' > /dev/null); then
+    if ! (cat tr.sfst | $PYTHON ./test_std_streams.py sfst 2 > tmp.sfst); then
+	echo "ERROR: in sfst"
+	rm tr.sfst
+	exit 1
+    fi
+    if ! ($COMPARE tr.sfst tmp.sfst); then
+	echo "ERROR in sfst"
+	rm tr.sfst tmp.sfst
+	exit 1
+    fi
     rm tr.sfst tmp.sfst
-    exit 1
 fi
-rm tr.sfst tmp.sfst
 
 # OPENFST
 if ! (cat tr.ofst | $PYTHON ./test_std_streams.py openfst 2 > tmp.ofst); then
@@ -75,14 +85,16 @@ fi
 rm tr.ofst tmp.ofst
 
 # FOMA
-if ! (cat tr.foma | $PYTHON ./test_std_streams.py foma 2 > tmp.foma); then
-    rm tr.foma
-    echo "ERROR: in foma"
-    exit 1
-fi  
-if ! ($COMPARE tr.foma tmp.foma); then
-    echo "ERROR: in foma"
+if (hfst-format --list | grep 'foma' > /dev/null); then
+    if ! (cat tr.foma | $PYTHON ./test_std_streams.py foma 2 > tmp.foma); then
+	rm tr.foma
+	echo "ERROR: in foma"
+	exit 1
+    fi  
+    if ! ($COMPARE tr.foma tmp.foma); then
+	echo "ERROR: in foma"
+	rm tr.foma tmp.foma
+	exit 1
+    fi  
     rm tr.foma tmp.foma
-    exit 1
-fi  
-rm tr.foma tmp.foma
+fi
